+++ src/test/java/ca/uhn/fhir/jpa/starter/CustomInterceptorTest.java
@@ -40,7 +40,7 @@ class CustomInterceptorTest {
 		ctx = FhirContext.forR4();
 		ctx.getRestfulClientFactory().setServerValidationMode(ServerValidationModeEnum.NEVER);
 		ctx.getRestfulClientFactory().setSocketTimeout(1200 * 1000);
-		String ourServerBase = "http://localhost:" + port + "/fhir/";
+		String ourServerBase = "http://localhost:" + port + "/fhir/DEFAULT/";
 		client = ctx.newRestfulGenericClient(ourServerBase);
 
 		// Properties props = new Properties();