+++ src/test/java/ca/uhn/fhir/jpa/starter/ParallelUpdatesVersionConflictTest.java
@@ -55,7 +55,7 @@ public class ParallelUpdatesVersionConflictTest {
 		ctx = FhirContext.forR4();
 		ctx.getRestfulClientFactory().setServerValidationMode(ServerValidationModeEnum.NEVER);
 		ctx.getRestfulClientFactory().setSocketTimeout(1200 * 1000);
-		String ourServerBase = "http://localhost:" + port + "/fhir/";
+		String ourServerBase = "http://localhost:" + port + "/fhir/DEFAULT/";
 		client = ctx.newRestfulGenericClient(ourServerBase);
 	}